DLUHC-201 move reusable inputs into seperate components
