+++ dluhc-component-library/src/components/timetableForm/components/TitlePage.tsx
@@ -1,4 +1,4 @@
-import Textarea from "./Textarea";
+import Textarea from "src/components/textarea/Textarea";
 
 const TitlePage = () => {
   return (
